==> notebooks/time/multi-modality/11_preprocessing/1_scMultiome_Mouse-Islets_NVF_E14-E16_Preprocessing_Split-and-DropletUtils_v2.ipynb <==
# NVF Mouse Islets - Preprocessing - Prepare Data for DropletUtils

2023-11-25 10:44:48    

# Setup

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:85% !important; }</style>"))

# General
import scipy as sci
import numpy as np
import pandas as pd
import logging
import gc # Free memory #gc.collect()

# Analysis
import scanpy as sc

# Settings

## Scanpy settings
sc.settings.verbosity = 3
sc.logging.print_versions()

# Setup R

#R
import os
os.environ['R_HOME'] = '/home/michi/Software/venvs/scverse_2023/lib/R' #path to your R installation

import rpy2
import rpy2.robjects as ro
import rpy2.rinterface_lib.callbacks
from rpy2.robjects import pandas2ri
import anndata2ri

## R settings

### Ignore R warning messages
#### Note: this can be commented out to get more verbose R output
rpy2.rinterface_lib.callbacks.logger.setLevel(logging.ERROR)

### Automatically convert rpy2 outputs to pandas dataframes
pandas2ri.activate()
anndata2ri.activate()
%load_ext rpy2.ipython

%%R

.libPaths()

%%R
# General
#library(tidyverse) # incl. ggplot2
library(rhdf5)
library(Matrix)
library(SingleCellExperiment)
library(DropletUtils)

# Parallelization
library(BiocParallel)
register(MulticoreParam(64, progressbar = TRUE))

library(future)
plan(multicore, workers = 64)
options(future.globals.maxSize = 100 * 1024 ^ 3) # for 50 Gb RAM
plan()

library(doParallel)
registerDoParallel(64)

sessionInfo()

# Functions

def run_dropletUtils(sample=None, # sample name fround in base_path e.g. '21L011182'
                     base_path=None, # path to cellranger out put folder containing different samples e.g. '/storage/scRNA-seq/scRNA-seq_iPSC_IIR-KO_S5_P21092B/data/cr_rev_7/cellranger/'
                     outs_path=None # folder name containing count matrices (include "/" before) e.g. '/count_matrices'
                    ):
    ro.globalenv['sample'] = sample
    ro.globalenv['base_path'] = base_path
    ro.globalenv['outs_path'] = outs_path + '/'
    ro.r('''
    path_to_adata = paste0(base_path, sample, outs_path, sample, "_raw_gex_bc_matrix.h5ad")

    print(paste0('Loading ',path_to_adata))

    adata <- h5read(path_to_adata, "/", compoundAsDataFrame=FALSE)

    barcodes <- adata$obs$`_index`
    genes <- adata$var$`_index`
    counts <- adata$X$data
    indices <- adata$X$indices
    pointer <- adata$X$indptr

    print("Construct SingleCellExperiment object as input for DropUtils")
    sparse_mat <- sparseMatrix(p = as.numeric(pointer), x= as.numeric(counts),  i = as.numeric(indices)+1)
    sce <- SingleCellExperiment(assays = list(counts = sparse_mat), colData=barcodes)
    rownames(sce) <- genes

    print("Computing barcode ranks")
    barcode_ranks <- barcodeRanks(counts(sce))

    print("Run DropUtils")
    drops <- emptyDrops(counts(sce))
    rownames(drops) <- colData(sce)$X
    is_cell <- drops$FDR <= 0.05

    print("Save output")
    cell_barcodes <- barcodes[which(is_cell)]
    write.csv(cell_barcodes, file=paste0(base_path, sample, outs_path, sample, "_DropletUtils_CellBarcodes.csv"))

    ambient_genes <- drops@metadata$ambient
    write.csv(data.frame(ambient_genes), file=paste0(base_path, sample, outs_path, sample, "_DropletUtils_AmbientGenes.csv"))

    cell_probs <- drops$LogProb
    write.csv(data.frame(barcodes=rownames(drops), cell_probs=cell_probs), file=paste0(base_path, sample, outs_path, sample, "_DropletUtils_LogProbabilities.csv"))
    ''')

# Prepare Data for DropletUtils

Load raw data filter out droplets with less than 1 counts and save.

samples = ['E14_5','E15_5','NVF_E15-5_Rep2','NVF_E16-5_Rep1']
base_path = '/storage/scRNA-seq/scMultiome_Mouse-Islets_NVF_E14-E16_P23044/cr_arc_rev9/cr_count/'
outs_path = '/outs'

for sample in samples:
    print('Loading ' + base_path + sample + outs_path)
    path = base_path + sample + outs_path
    adata = sc.read_10x_h5(path + '/raw_feature_bc_matrix.h5', gex_only=False)
    print(adata.shape)
    sc.pp.filter_cells(adata, min_counts=1)
    sc.pp.filter_genes(adata, min_cells=1)
    print(adata.shape,'\n\n')
    # Save combined
    sc.write(path + '/' + sample + '_raw_feature_bc_matrix', adata)
    # Split and save GEX and ATAC
    print('Shape GEX:', adata[:,adata.var.feature_types.isin(['Gene Expression'])].shape)
    sc.write(path + '/' + sample + '_raw_gex_bc_matrix', adata[:,adata.var.feature_types.isin(['Gene Expression'])])
    print('Shape ATAC:', adata[:,adata.var.feature_types.isin(['Peaks'])].shape)
    sc.write(path + '/' + sample + '_raw_atac_bc_matrix', adata[:,adata.var.feature_types.isin(['Peaks'])])
    del adata
    gc.collect()
    run_dropletUtils(sample=sample, base_path=base_path, outs_path=outs_path)
