==> notebooks/time/multi-modality/11_preprocessing/9-07_export_obs_names.ipynb <==
import muon

mudata = muon.read("/lustre/groups/ml01/workspace/moscot_paper/pancreas_revision/mudata_with_annotation_all.h5mu")

endocrine_celltypes = [
    "Ngn3 low",
    "Ngn3 high",
    "Ngn3 high cycling",
    "Fev+",
    "Fev+ Alpha",
    "Fev+ Beta",
    "Fev+ Delta",
    "Eps. progenitors",
    "Alpha",
    "Beta",
    "Delta",
    "Epsilon"
]

mudata.obs[["cell_type", "cell_type_refined"]].to_csv("/lustre/groups/ml01/workspace/moscot_paper/pancreas_revision/cell_type_annotation.csv")

mudata_filtered = mudata[mudata.obs["cell_type"].isin(endocrine_celltypes)]

mudata_filtered.obs[["cell_type", "cell_type_refined"]].to_csv("/lustre/groups/ml01/workspace/moscot_paper/pancreas_revision/cell_type_annotation_endocrine.csv")
