==> notebooks/time/multi-modality/11_preprocessing/9-08_filter_seurat.ipynb <==
library(anndata)
library(Signac)
library(Seurat)
library(tidyverse)

seurat <- readRDS("/lustre/groups/shared/users/idr_icb/scMultiome_Mouse-Islets_NVF_E14-E16/scMultiome_Mouse-Islets_NVF_E14-E16_seurat_markedDoublets_mergedPeaks_normalized.rds")

new_annotations <- read.csv("/lustre/groups/ml01/workspace/moscot_paper/pancreas_revision/cell_type_annotation.csv")

cells_to_keep <- colnames(seurat) %in% new_annotations$X

filtered_seurat_object <- seurat[, cells_to_keep]


filtered_seurat_object

DefaultAssay(object=filtered_seurat_object) <- "atac"

Fragments(filtered_seurat_object) <- NULL

fragments <- CreateFragmentObject(path = "/lustre/groups/shared/users/idr_icb/scMultiome_Mouse-Islets_NVF_E14-E16/atac_fragments.tsv.gz", cells = colnames(filtered_seurat_object), validate.fragments = TRUE)

Fragments(filtered_seurat_object) <- fragments

filtered_seurat_object$cell_type <- new_annotations$cell_type

filtered_seurat_object$cell_type_refined <- new_annotations$cell_type_refined

saveRDS(filtered_seurat_object, file="/lustre/groups/ml01/workspace/moscot_paper/pancreas_revision/seurat_annotated.rds")

new_annotations_endocrine <- read.csv("/lustre/groups/ml01/workspace/moscot_paper/pancreas_revision/cell_type_annotation_endocrine.csv")

cells_to_keep <- colnames(filtered_seurat_object) %in% new_annotations_endocrine$X

filtered_seurat_object_endocrine <- filtered_seurat_object[, cells_to_keep]

filtered_seurat_object_endocrine@meta.data <- cbind(filtered_seurat_object_endocrine@meta.data, new_annotations_endocrine)

saveRDS(filtered_seurat_object_endocrine, file="/lustre/groups/ml01/workspace/moscot_paper/pancreas_revision/seurat_endocrine_annotated.rds")
